requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0
scikit-image>=0.21.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Machine Learning
scikit-learn>=1.3.0
faiss-cpu>=1.7.4  # For efficient nearest neighbor search
# faiss-gpu>=1.7.4  # Uncomment if GPU available

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Metrics and Evaluation
torchmetrics>=1.0.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0
python-dotenv>=1.0.0

# Logging and Monitoring
tensorboard>=2.13.0
wandb>=0.15.0  # Optional: for experiment tracking
tqdm>=4.65.0

# Utilities
timm>=0.9.0  # Pre-trained models
einops>=0.6.1
kornia>=0.7.0  # Advanced augmentations

# Development
jupyter>=1.0.0
ipython>=8.14.0
ipywidgets>=8.0.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
